# Match kings() call more cleanly with other piece calls