--- a/src/chess/bitboard.h
+++ b/src/chess/bitboard.h
@@ -49,6 +49,7 @@ class BoardSquare {
   BoardSquare(const std::string& str, bool black = false)
       : BoardSquare(black ? '8' - str[1] : str[1] - '1', str[0] - 'a') {}
   constexpr std::uint8_t as_int() const { return square_; }
+  constexpr std::uint8_t as_board() const { return 1ULL << square_; }
   void set(int row, int col) { square_ = row * 8 + col; }
 
   // 0-based, bottom to top.
@@ -221,7 +222,7 @@ class BitBoard {
 
   // Returns bitboard with one bit reset.
   friend BitBoard operator-(const BitBoard& a, const BoardSquare& b) {
-    return {a.board_ & ~(1ULL << b.as_int())};
+    return {a.board_ & ~b.as_board()};
   }
 
   // Returns difference (bitwise AND-NOT) of two boards.