--- a/src/neural/encoder.cc
+++ b/src/neural/encoder.cc
@@ -111,14 +111,14 @@ InputPlanes EncodePositionForNN(
     result[base + 2].mask = (board.ours() & board.bishops()).as_int();
     result[base + 3].mask = (board.ours() & board.rooks()).as_int();
     result[base + 4].mask = (board.ours() & board.queens()).as_int();
-    result[base + 5].mask = (board.our_king()).as_int();
+    result[base + 5].mask = (board.ours() & board.kings()).as_int();
 
     result[base + 6].mask = (board.theirs() & board.pawns()).as_int();
     result[base + 7].mask = (board.their_knights()).as_int();
     result[base + 8].mask = (board.theirs() & board.bishops()).as_int();
     result[base + 9].mask = (board.theirs() & board.rooks()).as_int();
     result[base + 10].mask = (board.theirs() & board.queens()).as_int();
-    result[base + 11].mask = (board.their_king()).as_int();
+    result[base + 11].mask = (board.theirs() & board.kings()).as_int();
 
     const int repetitions = position.GetRepetitions();
     if (repetitions >= 1) result[base + 12].SetAll();