same idea as #1131

this also adds an `as_board()` method to BoardSquare to save us from writing `1ULL << *.as_int()` more than once

Benchmarks:
```
- Master
Benchmark final time 10.0009s calculating 212417 nodes per second.
Benchmark final time 10.0006s calculating 208707 nodes per second.
Benchmark final time 10.0004s calculating 208102 nodes per second.

- Knights
Benchmark final time 10.0005s calculating 208702 nodes per second.
Benchmark final time 10.0004s calculating 204465 nodes per second.
Benchmark final time 10.0006s calculating 206022 nodes per second.

- Kings
Benchmark final time 10.0005s calculating 206489 nodes per second.
Benchmark final time 10.0005s calculating 207881 nodes per second.
Benchmark final time 10.0006s calculating 206204 nodes per second.
```