@mooskagh 2020-03-18
Similarly to another PR, that's probably performance neutral, but as it merges positions of kings in to separate them back, would be nice to run a benchmark to check it.